+++ CHANGELOG.md
@@ -27,6 +27,8 @@
 * [Tmux] Use the color 231 instead of 15 for white to avoid issue with terminal
     theme switches
 * [Zsh/Aliases] Add `wnet` for `watch -n 1 nmcli`
+* [Zsh/Aliases] `gmod` and `gmodg` now work as expected with falsely execatables
+    files
 * [Zsh/Docker] Add start / stop aliases
 * [Zsh/Docker] Add `--sig-proxy=false` to the `docker attach` alias
 * [Zsh/Docker] Add volume management aliases